chore: interfaces in wrong folder
