+++ src/CitMovie.Business/ManagerInterfaces/IBookmarkManager.cs
