+++ src/CitMovie.Business/ManagerInterfaces/ICompletedManager.cs
