+++ src/CitMovie.Business/ManagerInterfaces/IPromotionalMediaManager.cs
